requirements.txt
paramiko
python-nmap
termcolor
datetime
argparse
